+++ core-domain/src/main/java/com/snapreceipt/io/domain/repository/ReceiptRepository.kt
@@ -5,11 +5,11 @@ import kotlinx.coroutines.flow.Flow
 
 interface ReceiptRepository {
     fun getAllReceipts(): Flow<List<ReceiptEntity>>
-    suspend fun getReceiptById(id: Int): ReceiptEntity?
+    suspend fun getReceiptById(id: Long): ReceiptEntity?
     suspend fun insertReceipt(receipt: ReceiptEntity): Long
     suspend fun updateReceipt(receipt: ReceiptEntity)
     suspend fun deleteReceipt(receipt: ReceiptEntity)
-    suspend fun deleteReceipts(ids: List<Int>)
+    suspend fun deleteReceipts(ids: List<Long>)
     fun getReceiptsByDateRange(startDate: Long, endDate: Long): Flow<List<ReceiptEntity>>
     fun getReceiptsByType(type: String): Flow<List<ReceiptEntity>>
 }
