+++ core-domain/src/main/java/com/snapreceipt/io/domain/usecase/receipt/DeleteReceiptsUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class DeleteReceiptsUseCase @Inject constructor(
     private val repository: ReceiptRepository
 ) {
-    suspend operator fun invoke(ids: List<Int>): Result<Unit> =
+    suspend operator fun invoke(ids: List<Long>): Result<Unit> =
         runCatching { repository.deleteReceipts(ids) }
 }
